object-engine: integrate filestore with master
