+++ src/object-engine/master/src/fs.rs
@@ -12,12 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod fs;
-mod master;
-pub mod proto;
-mod server;
+use std::{path::PathBuf, sync::Arc};
 
-use object_engine_common::{Error, Result};
+use object_engine_filestore::{local, Bucket, Store, Tenant};
 
-use self::fs::{FileBucket, FileStore, FileTenant};
-pub use self::{master::Master, server::Server};
+use crate::Result;
+
+pub type FileStore = Arc<dyn Store>;
+pub type FileTenant = Arc<dyn Tenant>;
+pub type FileBucket = Arc<dyn Bucket>;
+
+pub async fn open(path: impl Into<PathBuf>) -> Result<FileStore> {
+    let store = local::Store::open(path).await?;
+    let store: Box<dyn Store> = Box::new(store);
+    Ok(store.into())
+}
